feat: add trace id into log
